Add files via upload
New features like disposable income based on WI laws

In the State of Wisconsin, law limits the amount that can be garnished for a child support order as follows. 50% of disposable income if the payer has an intact family living with her or him (a spouse and/or child) and has no arrears 55% of disposable income if the payer has an intact family living with her or him (a spouse and/or child) and has arrears 60% of disposable income if the payer has no intact family (a spouse and/or child) living with her or him and has no arrears 65% of disposable income if the payer has no intact family (a spouse and/or child) living with her or him and has arrears The State of Wisconsin defines gross income as all of the employee’s income from all sources before mandatory deductions for federal, state, local, and Social Security taxes are deducted. Gross income also includes employee contributions to any employee benefit program or profit-sharing and voluntary contributions to any pension or retirement account whether or not the account provides for tax deferral or avoidance.
The State of Wisconsin defines disposable income as the part of the earnings of the employee remaining after deducting federal, state, and local withholding taxes, and Social Security taxes. Deductions for Individual Retirement Accounts, medical expense accounts, etc. do not reduce disposable income.